+++ src/main/java/com/example/numberbaseball/service/MessageService.java
@@ -1,0 +1,34 @@
+package com.example.numberbaseball.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.Nullable;
+import org.springframework.messaging.MessageHeaders;
+import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
+import org.springframework.messaging.simp.SimpMessageType;
+import org.springframework.messaging.simp.SimpMessagingTemplate;
+import org.springframework.stereotype.Service;
+
+import com.example.numberbaseball.Enum.RoomStatus;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Service
+public class MessageService {
+	@Autowired
+	private final SimpMessagingTemplate messagingTemplate;
+
+
+	public void message1(String receiverSessionId, RoomStatus roomStatus){
+		messagingTemplate.convertAndSendToUser(receiverSessionId,
+			"/topic/waiting/", roomStatus, createHeaders(receiverSessionId)
+		);
+	}
+
+	private MessageHeaders createHeaders(@Nullable String sessionId) {
+		SimpMessageHeaderAccessor headerAccessor = SimpMessageHeaderAccessor.create(SimpMessageType.MESSAGE);
+		if (sessionId != null) headerAccessor.setSessionId(sessionId);
+		headerAccessor.setLeaveMutable(true);
+		return headerAccessor.getMessageHeaders();
+	}
+}
