+++ src/main/resources/templates/mainPage.html
@@ -16,7 +16,8 @@
 <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.5.1/sockjs.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"></script>
 <script>
-    const host = "211.184.188.32"
+    // const host = "211.184.188.32"
+    const host = "localhost"
     const ports = "18082"
     var roomCode = null;
